remove old dependencies and update README
