+++ README.md
@@ -1,29 +1,27 @@
 Build: [![Build Status](https://travis-ci.org/hunterwerlla/podman.svg?branch=master)](https://travis-ci.org/hunterwerlla/podman)
 ### Podman - a terminal based podcast client
-Podman is a terminal based podcast client written in Go.
+Podman is a terminal based podcast client written in Go with a fully features TUI and somewhat functional CUI
 
-### Default Key Binds
-∧ to scroll up 
+### Default Key Bindings
+**K/&lt;Up&gt;** to scroll up<br/>
+**J/&lt;Down&gt;** to scroll down<br/>
+**H/&lt;**- to move left (search)<br/>
+**L/-&gt;** to move right (downloaded)<br/>
+**PgUp** to skip forward<br/>
+**PgDown** to skip backward<br/>
+**&lt;spacebar&gt;** to pause/resume when playing<br/>
+**&lt;enter&gt;** to do actions (play, download, view all of the episodes, subscribe)<br/>
+**/** to search podcast list/downloads/new podcasts to subscribe to
 
-∨ to scroll down
-
-&lt; to search
-
-&gt; to see downloaded
-
-PgUp to skip forward
-
-PgDown to skip backward
-
-&lt;spacebar&gt; to pause and resume
-
-&lt;enter&gt; to do actions (play, download, switch to detailed view, subscribe)
-
+Keybinds are editable by changing ~/.config/podman/config.json
 
 ### Dependencies
 Podman requires:
 
-[gocui](https://github.com/jroimartin/gocui) for the interface, [go-sox](https://github.com/krig/go-sox) to play audio files, [sanitize](https://github.com/kennygrant/sanitize), [go-rss](https://github.com/ungerik/go-rss), and [pb](https://github.com/cheggaaa/pb).
+[termui](https://github.com/gizak/termui) for the interface,  
+[go-sox](https://github.com/krig/go-sox) to play audio files,  
+[sanitize](https://github.com/kennygrant/sanitize) to clean up podcast descriptions,  
+and [go-rss](https://github.com/ungerik/go-rss) to grab podcast feeds. 
 
 Podman utilizes go modules to simplify this.
