+++ go.mod
@@ -1,21 +1,15 @@
 module github.com/hunterwerlla/podman
 
 require (
-	github.com/cheggaaa/pb v1.0.27+incompatible
 	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/fatih/color v1.7.0 // indirect
 	github.com/gizak/termui v0.0.0-20181203230344-3461d4a2e81a
-	github.com/jroimartin/gocui v0.4.0
 	github.com/kennygrant/sanitize v1.2.4
 	github.com/krig/go-sox v0.0.0-20180617124112-7d2f8ae31981
-	github.com/mattn/go-colorable v0.0.9 // indirect
-	github.com/mattn/go-isatty v0.0.4 // indirect
+	github.com/maruel/panicparse v1.1.1 // indirect
 	github.com/mattn/go-runewidth v0.0.4 // indirect
 	github.com/mitchellh/go-wordwrap v1.0.0 // indirect
 	github.com/nsf/termbox-go v0.0.0-20181027232701-60ab7e3d12ed // indirect
 	github.com/paulrosania/go-charset v0.0.0-20151028000031-621bb39fcc83 // indirect
 	github.com/ungerik/go-rss v0.0.0-20180730053843-e6dab863fc93
 	golang.org/x/net v0.0.0-20181220203305-927f97764cc3 // indirect
-	golang.org/x/sys v0.0.0-20181221143128-b4a75ba826a6 // indirect
-	gopkg.in/cheggaaa/pb.v1 v1.0.27 // indirect
 )
